fix: collapsed menu items
