+++ src/components/Header.jsx
@@ -1,15 +1,6 @@
-import { useLocation } from "react-router";
 import HeaderNav from "./Header-Nav";
 
 export default function Header() {
-  const location = useLocation().pathname;
-
-  const modes = [
-    { name: "Designer", url: "design" },
-    { name: "Developer", url: "" },
-    { name: "Dabbler", url: "dabbles" },
-  ];
-
   return (
     <header>
       <a href="/" className="logo icon">
@@ -20,17 +11,7 @@ export default function Header() {
         <p>hi</p>
       </div>
 
-      <ul className="mode-nav">
-        {modes.map((m) => (
-          <a key={m.name} href={`/${m.url}`}>
-            <li className={location === `/${m.url}` ? "selected-mode" : null}>
-              {m.name}
-            </li>
-          </a>
-        ))}
-      </ul>
-
-      <HeaderNav items={["About"]} />
+      <HeaderNav />
     </header>
   );
 }
