+++ src/pages/Home.jsx
@@ -5,7 +5,7 @@ import { dev_projects } from "../data/dev-projects";
 export default function Home() {
   return (
     <main className="home">
-      <Hero selected="Developer" />
+      <Hero />
       <ProjectsContainer content={dev_projects} />
     </main>
   );
